Crate pythagoras

Pythagorean functions for right triangles. Simple helper functions to calculate sides and angles.

The methods in this library use a common theme in which the components of right triangles are consistenly referred to as:

a: Opposite side, or the rise, using rise/run language

b: Adjacent side, or the run, using rise/run language

c: Hypotenuse, or the diagonal, using rise/run language

r: The angle in radians

Method signatures are formed by adding the inputs (from the list above) to the desired outputs. Thus, if you have a and b, and you need c, call let c = ab_c(&a,&b);

Note, if you have an angle in degrees, simply pass in angle.to_radians(). And, to convert radians to degrees, use radians.to_degrees().

#Example

use pythagoras::right_angle::*;
use pythagoras::*;

// Using the standard 3,4,5 right triangle
const A:f32 = 3.0;
const B:f32 = 4.0;
const C:f32 = 5.0;
const R: f32 = 0.6435011;

// Get the hypotenuse (c)
let c = ab_c(&A,&B);
assert_eq!(c, C);

// Now get the angle (r) in radians
let r = ab_r(&A,&B);
assert_eq!(r, R);

// Convert the angle to degrees (using std::f32::to_degrees)
let r_degrees = r.to_degrees();
assert_eq!(r_degrees.round(), 37.0);

// Using the angle (r) and one side, get another side
let b = rc_b(&r,&C);
assert_eq!(b.round(), B.round());

let a = rc_a(&r,&C);
assert_eq!(a.round(), A.round());

// Using the angle (r), and one side (C), get the remaining two sides
let (a,b) = rc_ab(&r,&C);
assert_eq!(a, A);
assert_eq!(b, B);

Re-exports

pub use right_angle::*;

Modules

right_angle

RightAngle is designed to enable completion of right angle data based on the provided input. It provides a single method, TryFrom<&RightAngleInput>. Fill in whaever data you have, and RightAngle will fill in the rest.

Functions

ab_c

Returns the hypotenuse (c) of a right triangle given the rise (a) and run (b).

ab_r

Returns the radians (r) given the opposite side (a) and adjacent side (b).

ac_b

Returns the adjacent side (b) given the hypotenuse (c) and opposite side (a).

ac_r

Returns radians (r) given the opposite side (a) and hypotenuse (c).

bc_a

Returns the opposite side (a) given the hypotenuse (c) and adjacent side (b).

bc_r

Returns the radians (r) given the adjacent side (b) and hypotenuse (c).

ra_b

Returns the adjacent side (b) given the radians (r) and the opposite side (a).

ra_bc

Given radians and the opposite (a), calculate the adjacent (c) and hypotenuse (c).

ra_c

Returns the hypotenuse (c) given the radians (r) and the opposite side (a).

rb_a

Returns the opposite side(a) given the radians (r) and the adjacent side (b).

rb_ac

Given radians and the adjacent (b), calculate the opposite (a) and hypotenuse (c).

rb_c

Returns the hypotenuse (c) given the radians (r) and the adjacent side (b).

rc_a

Given the radians and the hypotenuse (c), return the opposite side (a)

rc_ab

Given radians and the hypotenuse (c), calculate the opposite (a) and adjacent (b).

rc_b

Given the radians and the hypotenuse (c), return the adjacent side (b)