pythagoras

Module right_angle

RightAngle is designed to enable completion of right angle data based on the provided input. It provides a single method, TryFrom<&RightAngleInput>. Fill in whaever data you have, and RightAngle will fill in the rest.

RightAngle (and RightAngleInput) supports serde. So you can Deserialize RightAngleInput and Serialize RightAngle. Perfect for API applications!

Example

use pythagoras::right_angle::{RightAngle, RightAngleInput};

// Using the standard 3,4,5 right triangle
const A:f32 = 3.0;
const B:f32 = 4.0;
const C:f32 = 5.0;
const R: f32 = 0.6435011;

const RIGHT_ANGLE: RightAngle = RightAngle {
    rise: A,
    run: B,
    diagonal: C,
    radians: R,
};

// Use one side and the angle to compete the rest of the right angle
let input = RightAngleInput {
    radians: Some(R),
    rise: Some(A),
    run: None,
    diagonal: None,
};

let result = RightAngle::try_from(&input).expect("Failed to create RightAngle");
assert_eq!(result, RIGHT_ANGLE);

// Use two sides to complete the rest of the right angle
let input = RightAngleInput {
   radians: None,
   rise: None,
   run: Some(B),
   diagonal: Some(C),
};

let result = RightAngle::try_from(&input).expect("Failed to create RightAngle");
assert_eq!(result, RIGHT_ANGLE);

// Create [RightAngle] by serializing [RightAngleInput] to
// a json string,
 let json = format!(r#"{{ "rise": {}, "run": {} }}"#, A, B);
 let result = RightAngle::try_from(json.as_str()).expect("Failed to convert");
 assert_eq!(result, RIGHT_ANGLE);

Structs

RightAngle

RightAngleInput

Functions

one_side

Given the angle (in radians) and one side, calculate the other two sides. Note, the first side found will be used, in order of rise (a), run (b), diagonal (c). An error is returned if no side is provided or the ange is not provided.

two_sides

Given two sides, calculate the third side. The angle is always calculated from the two sides given An error is returned if not enough sides are provided.